optimizations

+++ src/renderer/model.h
@@ -185,7 +185,7 @@ struct LUMIX_RENDERER_API Model final : Resource {
 	int getMeshCount() const { return m_meshes.size(); }
 	int getBoneCount() const { return m_bones.size(); }
 	const char* getBoneName(u32 idx) { return m_bones[idx].name.c_str(); }
-	i32 getBoneParent(u32 idx) const { return m_parents[idx]; }
+	i16 getBoneParent(u32 idx) const { return m_parents[idx]; }
 	const Bone& getBone(u32 i) const { return m_bones[i]; }
 	int getFirstNonrootBoneIndex() const { return m_first_nonroot_bone_index; }
 	BoneMap::ConstIterator getBoneIndex(BoneNameHash hash) const { return m_bone_map.find(hash); }
@@ -204,7 +204,7 @@ struct LUMIX_RENDERER_API Model final : Resource {
 	BoneNameHash getRootMotionBone() const { return m_root_motion_bone; }
 	const SOATransform& getInverseBindPose() { return m_inverse_bind; }
 	LocalRigidTransform getInverseBindTransform(i32 bone_idx) const;
-	Span<const i32> getParents() const { return m_parents; }
+	Span<const i16> getParents() const { return m_parents; }
 
 	static constexpr u32 MAX_LOD_COUNT = 4;
 
@@ -227,7 +227,7 @@ private:
 	Array<Bone> m_bones;
 
 	SOATransform m_inverse_bind;
-	Array<i32> m_parents;
+	Array<i16> m_parents;
 	LODMeshIndices m_lod_indices[MAX_LOD_COUNT + 1];
 	float m_lod_distances[MAX_LOD_COUNT];
 	float m_origin_bounding_radius = 0;
